Remove missing protocol stubs and fix nissan sign-compare
- protocol_items.h/c: drop honeywell, elplast, treadmill37, beninca_arc,
  jarolift, ditec_gol4, keyfinder — these are ARF-fork-only protocols
  with no source files in the official firmware base we build on
- nissan.c: change loop var to size_t for NISSAN_PREAMBLE_MIN comparison
  (NISSAN_PREAMBLE_MIN is unsigned; int vs unsigned triggers -Werror=sign-compare)

+++ new_files/nissan.c
@@ -191,7 +191,7 @@ static bool nissan_encoder_get_upload(SubGhzProtocolEncoderNissan* instance) {
     frame[7] = nissan_crc8(frame, 7);
 
     size_t idx = 0;
-    for(int i = 0; i < NISSAN_PREAMBLE_MIN; i++) {
+    for(size_t i = 0; i < NISSAN_PREAMBLE_MIN; i++) {
         instance->encoder.upload[idx++] =
             level_duration_make(true, subghz_protocol_nissan_const.te_short);
         instance->encoder.upload[idx++] =
